Add referrals, update webpack, and refactor components
