+++ public/dashboard.css
@@ -63,3 +63,31 @@ h1 {
 #stakeDataButton:hover {
     background-color: #367B35; /* A darker green on hover */
 }
+
+.header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+}
+
+.header-nav {
+    display: flex;
+    align-items: center;
+    gap: 20px; /* Adjust the gap between navigation links as needed */
+}
+
+.nav-link {
+    text-decoration: none;
+    color: #333;
+    font-size: 16px;
+    padding: 5px 10px;
+    border-radius: 5px;
+    transition: background-color 0.3s, color 0.3s;
+}
+
+.nav-link:hover,
+.nav-link.active {
+    background-color: #008CBA;
+    color: white;
+}
