+++ public/referrals.html
@@ -1,10 +1,10 @@
-<!-- public/dashboard.html -->
+<!-- public/referrals.html -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Dashboard</title>
-    <link rel="stylesheet" href="dashboard.css">
+    <title>Referrals</title>
+    <link rel="stylesheet" href="dashboard.css"> <!-- Assuming you want to use the same stylesheet -->
 </head>
 <body>
     <div class="header">
@@ -17,7 +17,10 @@
             <button id="walletButton">Connect Wallet</button>
         </div>
     </div>
-    <!-- Add dashboard content here -->
-    <script src="dashboard.js"></script>
+    <div class="content">
+        <!-- Content for referrals page goes here -->
+        <p>Welcome to the Referrals page. Share your referral link and earn rewards!</p>
+    </div>
+    <script src="referrals.js"></script> <!-- Link to the referrals TypeScript compiled output -->
 </body>
 </html>
